Project Creation Technologies
This project interesting movie names with the names of lead actor, actress, year of release and the director name is created by using the HTML, CSS, PHP and MySQL as a Database. You will see the two files above one is the actual code and other is the MySQL database.